refactor: prestringutil -> internal.prestringutil
